% Auto-generated: do not edit by hand
\name{'hg'HorizonGraph}

\alias{'hg'HorizonGraph}

\title{HorizonGraph component}

\description{
ExampleComponent is an example component. It takes a property, `label`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
'hg'HorizonGraph(id=NULL, bands=NULL, bgcolor=NULL, fontSize=NULL,
height=NULL, legendOffset=NULL, names=NULL, numBrushes=NULL,
numberHorizons=NULL, padding=NULL, redraw=NULL,
selectedRanges=NULL, threshold=NULL, width=NULL, x=NULL,
xLabel=NULL, y=NULL, yLabel=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{bands}{Numeric. }

\item{bgcolor}{Character. }

\item{fontSize}{Numeric. }

\item{height}{Numeric. }

\item{legendOffset}{Numeric. }

\item{names}{Unnamed list. }

\item{numBrushes}{Numeric. }

\item{numberHorizons}{Numeric. }

\item{padding}{Numeric. }

\item{redraw}{Logical. }

\item{selectedRanges}{Unnamed list. }

\item{threshold}{Numeric. }

\item{width}{Numeric. }

\item{x}{Unnamed list. }

\item{xLabel}{Character. }

\item{y}{Unnamed list. }

\item{yLabel}{Character. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

